removed types file, moved types closer to code, and split code
